IReadOnlyEntityPersistentData - Update docs and separate file
