+++ Scripts/Runtime/Entities/PersistentData/Data/Interface/IReadOnlyEntityPersistentData.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 3f070b847f1444098058efe0666e2f80
+timeCreated: 1681347335
